# Performance test for comparing PAIR sockets vs BlockingCollection