--- a/src/NetMQ.v11.sln
+++ b/src/NetMQ.v11.sln
@@ -1,5 +1,5 @@
 ﻿
-Microsoft Visual Studio Solution File, Format Version 11.00
+Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 2012
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NetMQ", "NetMQ\NetMQ.csproj", "{82934BAC-07FB-41AC-AE59-46FEE6026A40}"
 EndProject
@@ -26,6 +26,8 @@ EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WeatherUpdateClient", "Samples\WeatherUpdater\WeatherUpdateClient\WeatherUpdateClient.csproj", "{37B320C0-9E44-44EF-B89A-E671DBA4986B}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WeatherUpdateServer", "Samples\WeatherUpdater\WeatherUpdateServer\WeatherUpdateServer.csproj", "{5B8F7740-EFD5-46B9-8A96-2E1F47385295}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NetMQ.PerformanceComparison", "Performance\PerformanceComparison\NetMQ.PerformanceComparison.csproj", "{194F29F8-FBE2-4F35-88F3-9283C72B18F3}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -147,6 +149,18 @@ Global
 		{5B8F7740-EFD5-46B9-8A96-2E1F47385295}.Release|x64.ActiveCfg = Release|x86
 		{5B8F7740-EFD5-46B9-8A96-2E1F47385295}.Release|x86.ActiveCfg = Release|x86
 		{5B8F7740-EFD5-46B9-8A96-2E1F47385295}.Release|x86.Build.0 = Release|x86
+		{194F29F8-FBE2-4F35-88F3-9283C72B18F3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{194F29F8-FBE2-4F35-88F3-9283C72B18F3}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{194F29F8-FBE2-4F35-88F3-9283C72B18F3}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
+		{194F29F8-FBE2-4F35-88F3-9283C72B18F3}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
+		{194F29F8-FBE2-4F35-88F3-9283C72B18F3}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{194F29F8-FBE2-4F35-88F3-9283C72B18F3}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{194F29F8-FBE2-4F35-88F3-9283C72B18F3}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{194F29F8-FBE2-4F35-88F3-9283C72B18F3}.Release|Any CPU.Build.0 = Release|Any CPU
+		{194F29F8-FBE2-4F35-88F3-9283C72B18F3}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
+		{194F29F8-FBE2-4F35-88F3-9283C72B18F3}.Release|Mixed Platforms.Build.0 = Release|Any CPU
+		{194F29F8-FBE2-4F35-88F3-9283C72B18F3}.Release|x64.ActiveCfg = Release|Any CPU
+		{194F29F8-FBE2-4F35-88F3-9283C72B18F3}.Release|x86.ActiveCfg = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -156,6 +170,7 @@ Global
 		{BFCBB08C-0DFE-4D8A-8288-E98E22D1FD24} = {1D49A4C6-C03C-4F72-9A27-42C8E5219BBA}
 		{6D5192C6-8E40-492C-B4B8-E1A325BA35F7} = {1D49A4C6-C03C-4F72-9A27-42C8E5219BBA}
 		{4BB8D1CB-12E1-4C88-ABE0-29358CD02101} = {1D49A4C6-C03C-4F72-9A27-42C8E5219BBA}
+		{194F29F8-FBE2-4F35-88F3-9283C72B18F3} = {1D49A4C6-C03C-4F72-9A27-42C8E5219BBA}
 		{C6CA5118-649F-4389-A36E-70F7600FB99B} = {0D41951D-83C3-45B0-B319-A424EBF530E2}
 		{5A526E6B-083E-419A-A2BD-A0C807E6FA0E} = {0D41951D-83C3-45B0-B319-A424EBF530E2}
 		{37B320C0-9E44-44EF-B89A-E671DBA4986B} = {C6CA5118-649F-4389-A36E-70F7600FB99B}