--- a/src/Performance/PerformanceComparison/NetMQ.PerformanceComparison.csproj
+++ b/src/Performance/PerformanceComparison/NetMQ.PerformanceComparison.csproj
@@ -9,8 +9,9 @@
     <AppDesignerFolder>Properties</AppDesignerFolder>
     <RootNamespace>NetMQ.PerformanceComparison</RootNamespace>
     <AssemblyName>NetMQ.PerformanceComparison</AssemblyName>
-    <TargetFrameworkVersion>v4.5</TargetFrameworkVersion>
+    <TargetFrameworkVersion>v4.0</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
+    <TargetFrameworkProfile />
   </PropertyGroup>
   <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
     <DebugSymbols>true</DebugSymbols>