--- a/src/Performance/PerformanceComparison/packages.config
+++ b/src/Performance/PerformanceComparison/packages.config
@@ -1,0 +1,4 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="NUnit" version="2.6.2" targetFramework="net45" />
+</packages>