--- a/src/packages/repositories.config
+++ b/src/packages/repositories.config
@@ -1,0 +1,4 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<repositories>
+  <repository path="..\Performance\PerformanceComparison\packages.config" />
+</repositories>