… src/NetMQ.sln
… src/NetMQ.vs10.sln
… src/packages/NUnit.2.6.2/NUnit.2.6.2.nupkg
… src/packages/NUnit.2.6.2/lib/nunit.framework.dll
… src/packages/NUnit.2.6.2/lib/nunit.framework.xml